Implementing error handling for ticket

+++ app/Http/Controllers/TicketController.php
@@ -10,6 +10,11 @@ use Illuminate\Support\Facades\Auth;
 class TicketController extends Controller
 {
     //
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function index()
     {
         $data = [];
@@ -21,17 +26,28 @@ class TicketController extends Controller
     {
         $user = Auth::user();
         $eventPrice = EventPrice::find($id);
+        if (!$eventPrice)
+            return abort(500);
+        if ($user->hasEvent($eventPrice->event_id))
+            return abort(500);
         $ticket = new Ticket;
         $ticket->user()->associate($user);
         $ticket->eventPrice()->associate($eventPrice);
         $ticket->save();
-        return redirect(route('ticket'));
+        if (Ticket::where('event_price_id', $eventPrice->id)->count()>$eventPrice->site->capacity)
+        {
+            $ticket->delete();
+            return abort(500);
+        }
+        return redirect()->route('ticket');
     }
 
     public function cancel($id)
     {
         $ticket = Ticket::find($id);
+        if (!$ticket)
+            return abort(500);
         $ticket->delete();
-        return redirect(route('ticket'));
+        return redirect()->route('ticket');
     }
 }
